BR-44: Move utility classes to the util package
